requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile setup.cfg
#
paho-mqtt==2.1.0
    # via zaphod_power (setup.cfg)
